+++ src/constants.ts
@@ -147,6 +147,7 @@ export const PACKAGES_TO_GET_INFO_FOR = [
   ...CHECKED_LODASH_METHODS.map((method) => `lodash.${method}` as const),
   'rxjs',
   'nx',
+  'zod',
 
   'astro',
   'vue',