+++ CHANGELOG.md
@@ -35,6 +35,7 @@
 - `headers` via [`eslint-plugin-headers`](https://npmjs.com/eslint-plugin-headers), **<u>disabled</u>** by default.
 - `rxjs` via [`@smarttools/eslint-plugin-rxjs`](https://npmjs.com/@smarttools/eslint-plugin-rxjs), enabled automatically if `rxjs` package is installed.
 - `nx` via [`@nx/eslint-plugin`](https://npmjs.com/@nx/eslint-plugin), enabled automatically if `nx` package is installed.
+- `importZod` via [`eslint-plugin-import-zod`](https://npmjs.com/eslint-plugin-import-zod), enabled automatically if both `zod` and `next` packages are installed.
 
 ### Changes